% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sphere.R
\name{add_sphere}
\alias{add_sphere}
\title{Add a 2D Shaded Sphere to a Plot}
\usage{
add_sphere(
  x0 = 0,
  y0 = 0,
  r0 = 1,
  x1 = -0.5,
  y1 = 0.5,
  r1 = 0.1,
  h = 0.01,
  n = 1000,
  col = c("black", "white")
)
}
\arguments{
\item{x0}{coordinate on the x-axis of center of the sphere}

\item{y0}{coordinate on the y-axis of center of the sphere}

\item{r0}{radius of the sphere}

\item{x1}{coordinate on the x-axis of center of the shade}

\item{y1}{coordinate on the y-axis of center of the shade}

\item{r1}{radius of the smallest shade}

\item{h}{factor of shading}

\item{n}{number of vertices of the circle}

\item{col}{vector of two colors used to generate a continuous gradient}
}
\description{
\strong{Add a Description}
}
\examples{
\dontrun{
plot(0, 0, asp = 1, xlim = c(-1, 1), ylim = c(-1, 1))
add_sphere()
}
}
\author{
Nicolas Casajus, \email{nicolas.casajus@gmail.com}
}
