% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_coords.R
\name{polar_coords}
\alias{polar_coords}
\title{Get Polar Coordinates of a Circle}
\usage{
polar_coords(x, y, r = 1, n = 1000)
}
\arguments{
\item{x}{coordinate on the x-axis of center of the circle}

\item{y}{coordinate on the y-axis of center of the circle}

\item{r}{radius of the circle}

\item{n}{number of vertices of the circle}
}
\value{
A two-columns data frame with polar coordinates of the circle.
}
\description{
\strong{Add a Description}
}
\examples{
\dontrun{
polar_coords(0, 0, 1, 500)
}
}
\author{
Nicolas Casajus, \email{nicolas.casajus@gmail.com}
}
